requirements.txt
Flask
flask-socketio
eventlet
paho-mqtt